admin_tool/requirements.txt
alembic==0.8.10
blinker==1.3
click==6.7
email==4.0.2
Flask==0.12
Flask-Admin==1.5.0
Flask-Bootstrap==3.0.3.1
Flask-Login==0.3.1
Flask-Mail==0.9.0
Flask-Migrate==2.0.3
Flask-Moment==0.2.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gevent==1.2.2
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
mysql==0.0.1
MySQL-python==1.2.5
pkg-resources==0.0.0
protobuf==3.3.0
python-editor==1.0.3
redis==2.10.5
six==1.10.0
SQLAlchemy==1.1.5
Werkzeug==0.11.15
WTForms==2.1
